Refactor many of the default package, charlm, and pretrain knowledge into a separate file.  Will avoid circular imports if something imported by stanza (the root package) needs to import the constants.  One example would be a model trying to find its own default charlm
